Moving a tank

+++ point.h
@@ -1,6 +1,6 @@
 #pragma once
 
 struct Point {
-  float x;
-  float y;
+  double x;
+  double y;
 };
